Partial implementation of get() functions
There are some things here that don't work very neatly:

(1) I have had to add default constructors to Scan and its subclasses
so that the JSON can deserialise the MongoDB document.

(2) I have also added a setId() that takes a hex string. This allows me
to convert the ObjectId (_id) to a GDA id in String format in the document.
This is easier than writing a codec for BigInteger.

(3) Type hierarchy. This is the big unsolved problem at the moment.
Currently, we put documents in a collection with the name of the class
immediately below PersistableItem, so that DiffractionScan and
TomographyScan both go in collection db.uk.ac.diamond.daq.persistence.data.Scan.
The actual class name is stored in the document, allowing us to tell
the ObjectMapper what sort of object to create.

The problem is searching. It is easy to return objects of a particular class
(e.g. DiffractionScan), but what if (as in the tests), we search for a superclass?
If we search for Scan objects, MongoDB doesn't know that that means
(TomographyScan + DiffractionScan). It gets even worse if there are more
levels of subclass e.g. TomographyStepScan & TomographyFlyScan.

Java doesn't really help here, as you can only go up the class
hierarchy, not down. I suppose one solution might be to store in each
document not only the actual class, but also the names of all superclasses.
It's not very scalable, though.